refactor(gsd): extract auto-worktree sync module
